feat: migrate CreateTrain screens to :feature:createTrain KMP module
- Create :feature:createTrain KMP module with 18 commonMain files
- Extract UiState classes from ViewModels into standalone files
- Consolidate duplicate TimeWheelPicker into single file
- Move all 4 step stateless screens + parent screen + shared composables
- Replace HapticFeedbackType.SegmentFrequentTick with LongPress (CMP compat)
- Delete 11 old files from :app, update 8 files to thin wrappers
