+++ feature/createTrain/src/commonMain/kotlin/com/sedilant/yambol/feature/createTrain/stepsScreens/concepts/Concept.kt
@@ -1,0 +1,7 @@
+package com.sedilant.yambol.feature.createTrain.stepsScreens.concepts
+
+data class Concept(
+    val id: String,
+    val conceptName: String,
+    val isSelected: Boolean = false
+)
