+++ feature/createTrain/src/commonMain/kotlin/com/sedilant/yambol/feature/createTrain/stepsScreens/concepts/ConceptsUiState.kt
@@ -1,0 +1,7 @@
+package com.sedilant.yambol.feature.createTrain.stepsScreens.concepts
+
+sealed interface ConceptsUiState {
+    data class Success(val concepts: List<Concept>) : ConceptsUiState
+    data class Error(val message: String) : ConceptsUiState
+    data object Loading : ConceptsUiState
+}
